requirements.txt
django
djangorestframework
redis
pyjwt
django-redis